Class game - deleted it I like GamePanel more as a name :).
Class GamePanel - I created a FIXED game loop, and also rendered the
window the game will be played on. Update and Input have not been
implemented but are stubs so basically i made a box that renders a blue
back round... Cool huh?

Class main - makes an object Window.

Class Window - makes and object GamePanel and set up the init. of
Creating the frame.